refactor: Update MathJax configuration to load additional packages
Update the MathJax configuration in constants.ts to load the "[tex]/unicode" package in addition to the existing "[tex]/html" package. This change allows for rendering mathematical expressions using Unicode characters.

Also, update the WithMathJax component in common/WithMathJax.tsx to use the children prop instead of the text prop. This change ensures that the component can render any content wrapped in MathJax.